Update softsplat_torch.py
Former-commit-id: 7c8757c0ad7181c4cbec49091db701726af85227
Former-commit-id: 2e03362d889f4d68098e55725d96e3990d8d7d92

+++ backend/src/pytorch/InterpolateArchs/util/softsplat_torch.py
@@ -38,7 +38,7 @@ def forward(tenIn, tenFlow):
     gridY,gridX = (gridY.unsqueeze(0).unsqueeze(0).expand(N, 1, H, W), gridX.unsqueeze(0).unsqueeze(0).expand(N, 1, H, W))
     
 
-    batch_indices = torch.arange(N, device=device).view(N, 1, 1).expand(N, H, W).reshape(-1)
+    batch_indices = torch.arange(N, device=device,dtype=origdtype).view(N, 1, 1).expand(N, H, W).reshape(-1)
